Allow logged in players to join room with current avatar
